+++ logic/main.go
@@ -16,3 +16,20 @@ func nor(a bool, b bool) bool {
 func xor(a bool, b bool) bool {
 	return (a || b) && !(a && b)
 }
+func not(a bool) bool {
+	return !a
+}
+func impl(a bool, b bool) bool {
+	return or(not(a), b)
+}
+func equ(a bool, b bool) bool {
+	return a == b
+}
+func printLogicTable(f func(bool, bool) bool) [4][3]bool {
+	return [4][3]bool {
+	   {true, true, f(true, true)},
+	   {true, false, f(true, false)},
+	   {false, true, f(false, true)},
+	   {false, false, f(false, false)},  
+	}
+}