+++ logic/p46_test.go
@@ -14,7 +14,7 @@ func TestAnd(t *testing.T) {
 		}
 	})
 	t.Run("false/true", func(t *testing.T) {
-		if and(true, false) {
+		if and(false, true) {
 			t.Error("Expected to be false")
 		}
 	})
@@ -37,7 +37,7 @@ func TestOr(t *testing.T) {
 		}
 	})
 	t.Run("false/true", func(t *testing.T) {
-		if !or(true, false) {
+		if !or(false, true) {
 			t.Error("Expected to be true")
 		}
 	})
@@ -60,7 +60,7 @@ func TestNand(t *testing.T) {
 		}
 	})
 	t.Run("false/true", func(t *testing.T) {
-		if !nand(true, false) {
+		if !nand(false, true) {
 			t.Error("Expected to be true")
 		}
 	})
@@ -83,7 +83,7 @@ func TestNor(t *testing.T) {
 		}
 	})
 	t.Run("false/true", func(t *testing.T) {
-		if nor(true, false) {
+		if nor(false, true) {
 			t.Error("Expected to be false")
 		}
 	})
@@ -106,7 +106,7 @@ func TestXor(t *testing.T) {
 		}
 	})
 	t.Run("false/true", func(t *testing.T) {
-		if !xor(true, false) {
+		if !xor(false, true) {
 			t.Error("Expected to be true")
 		}
 	})
@@ -116,3 +116,78 @@ func TestXor(t *testing.T) {
 		}
 	})
 }
+
+func TestImpl(t *testing.T) {
+	t.Run("all true", func(t *testing.T) {
+		if !impl(true, true) {
+			t.Error("Expected to be true")
+		}
+	})
+	t.Run("true/false", func(t *testing.T) {
+		if impl(true, false) {
+			t.Error("Expected to be false")
+		}
+	})
+	t.Run("false/true", func(t *testing.T) {
+		if !impl(false, true) {
+			t.Error("Expected to be true")
+		}
+	})
+	t.Run("all false", func(t *testing.T) {
+		if !impl(false, false) {
+			t.Error("Expected to be true")
+		}
+	})
+}
+
+func TestEqu(t *testing.T) {
+	t.Run("all true", func(t *testing.T) {
+		if !equ(true, true) {
+			t.Error("Expected to be true")
+		}
+	})
+	t.Run("true/false", func(t *testing.T) {
+		if equ(true, false) {
+			t.Error("Expected to be false")
+		}
+	})
+	t.Run("false/true", func(t *testing.T) {
+		if equ(false, true) {
+			t.Error("Expected to be false")
+		}
+	})
+	t.Run("all false", func(t *testing.T) {
+		if !equ(false, false) {
+			t.Error("Expected to be true")
+		}
+	})
+}
+
+func TestPrintLogicTable(t *testing.T) {
+	t.Run("print: and", func(t *testing.T) {
+		expectedResult := [4][3]bool{  
+		   {true, true, true},
+		   {true, false, false},
+		   {false, true, false},
+		   {false, false, false},  
+		}
+		result := printLogicTable(and)
+		if result != expectedResult {
+			t.Errorf("Expected to be same array. Expected %v, current: %v", expectedResult, result)
+		}
+	})
+	t.Run("print: impl", func(t *testing.T) {
+		expectedResult := [4][3]bool{  
+		   {true, true, true},
+		   {true, false, false},
+		   {false, true, true},
+		   {false, false, true},  
+		}
+		result := printLogicTable(impl)
+		if result != expectedResult {
+			t.Errorf("Expected to be same array. Expected %v, current: %v", expectedResult, result)
+		}
+	})
+}
+
+